+++ CHANGES.rst
@@ -5,7 +5,8 @@ Changelog
 0.2 (unreleased)
 ----------------
 
-- Nothing changed yet.
+- Fixed shift of PST structure in detailed description
+  [fngaha]
 
 
 0.1 (2019-06-23)